Use args argument instead of shlex.join in Action.run_cmd
The pexpect.spawn does not actually run commands in the shell.
Instead it actually has a built-in shlex.split functionality.

https://github.com/pexpect/pexpect/blob/aa989594e1e413f45c18b26ded1783f7d5990fe5/pexpect/utils.py#L69

When `args` argument is not passed to the pexpect.spawn it will
think that the `command` argument contains the shell command
and will split the arguments using its built-in splitter.

However, the Action.run_cmd already has a list of arguments so
instead of quoting and splitting back and forth simply pass the
args argument which should be all arguments except first one which
should be passed to the `command` argument.

To split the list of arguments in to a command and arguments use
the PEP 3132 Extended Iterable Unpacking. (`x, *y = [1, 2, 3]`)
This will separate the first item in a variable and put the rest
in to the list. In case when there is only one item the second
variable will become empty list.