Refactor __set_device_health__ function
That function was supposed to be used in both lava_scheduler_app
views and REST API v0.2. However, because of of differences between
django and restframework it did not work very well. For example,
the restframework uses a special Response class but __set_device_health__
returned the HttpResponce so the REST API code had to pick apart
the returned response and convert it to Response object.

Instead add a new method Device.set_health() that replaces the
__set_device_health__ function.

The .set_health() will check the permissions of the passed user
and raise the PermissionDenied exception that both django and
restframework can understand. The REST API v0.2 actually did not
check the device type permissions but Device.can_change() used
inside the .set_health() does.