Log command list directly in Action.run_cmd
Because since 7f072e16b612560377764b378f65070a99fe8f9c the commands
are directly passed to pexpect.spawn without joining them into
a single string it is more accurate to log the list itself rather
than joining all strings in to a pseudo shell command.

This also better logs the unprintable characters which will now be
escaped in to a printable representation.

Also only format the error message if an error occurred and
`error_msg` argument was None.