Revert "Use args argument instead of shlex.join in Action.run_cmd"
This reverts commit 7f072e16b612560377764b378f65070a99fe8f9c.

This is breaking meta-lava tests where a command is a single binary with spaces
in it. See https://gitlab.com/lava/meta-lava/-/jobs/6239072770#L629

LAVA logs:
```
Calling: ['/bin/with spaces']
Unable to run: The command was not found or was not executable: /bin/with.
Unable to run: ['/bin/with spaces']
```

Device dictionary snippet:
```
{% set flasher_deploy_commands = ["'/bin/with spaces'"] %}
```

This is due to a potential bug in pexpect where spawn is using [] as default
value for the command arguments. pexpect is then unable to know that args is
set to [] on purpose and force the parsing of the command as a string.
See https://github.com/pexpect/pexpect/blob/aa989594e1e413f45c18b26ded1783f7d5990fe5/pexpect/pty_spawn.py#L265

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>