Use pexpect's PopenSpawn for Action.run_cmd
The current CommandLogger used for pexpect.spawn does not account
for line mangling that pseudo TTY does. The PopenSpawn has the API
equivalent to pexpect.spawn but does not run it in a pseudo TTY.
This will prevent line mangling and make CommandLogger work as
intended.

Also PopenSpawn takes a list of arguments and does not have issues
with distinguishing between empty arguments and shell command
unlike the approach used in 7f072e16b612560377764b378f65070a99fe8f9c
which had to be reverted as pexpect.spawn would try to split the
single executable name with a spaces inside.

See PopenSpawn args handling:

https://github.com/pexpect/pexpect/blob/aa989594e1e413f45c18b26ded1783f7d5990fe5/pexpect/popen_spawn.py#L50