tests: Do not require `/dev/kvm` to run TestQemuNFS test
Only the presence of `/dev/kvm` is checked but it is not actually
used. This caused the TestQemuNFS to be skipped in the CI because
CI runs in a container with no KVM. Patch the variable containing
the `/dev/kvm` to `/` which should always pass the check.

Also drop the Python 3.5 check because the minimum Python version
right now is 3.9.