Improve coverage integration with Gitlab
Right now the coverage stage only reports the total percent
of coverage. However, it is possible to report lines of code
where to the coverage was changed each MR. To do this a new XML
report has to be uploaded as an artifact.

The example from Gitlab's documentation demonstrates the pytest-cov
integration with Gitlab:

https://docs.gitlab.com/ee/ci/testing/test_coverage_visualization.html#python-example

This should allow Gitlab to display the exact lines the coverage
changed. Set the `source` parameter for coverage.py to `.` (current directory)
as Gitlab requires the paths in XML report to be relative to
the project's root.

Also use the regex Gitlab documentation provides for the total
percent coverage.