Implement dispatcher's device model instead of untyped dict
Untyped dict is a source of many errors because there is no way
to check if the used key and value are correct.

This commit implements the lava_dispatcher's Device model using
Pydantic. It can work alongside the current voluptuous schema.
The voluptuous will check the schema and Pydantic will convert
the dictionary to objects.

Also add Pydantic as the requirements for both server and dispatcher
as in the future it will replace voluptuous.