tests: Let create_kvm_job raise errors instead of returning None
For some reason Factory.create_kvm_job would catch exceptions
raised by job parser and return None instead. This is counter-productive
as most likely an AttributeError would be raised next as the None
would be attempted to be used as Job.

Instead just let it raise errors as that will show the exact error
raised by job parser in the output.