Make Action.job always be a valid job
Currently the `.job` attribute is initialized to None in Action.__init__
and is set to a Job during the Pipeline.add_action. However, this
causes issues for type checking because you would need to check for
None every time the `.job` attribute is accessed.

Instead require that Action.__init__ is called with a valid Job
as the first argument. This requires a lot of changes everywhere
in the code and unit tests but it is a more correct approach and
will enable strict type checking in the future.