Log pexpect input in to a input logs instead of mixing it with output logs
By default pexpect logs all input and output in to a single passed
`logfile`. However, this results that both input and output are
mixed in to a single log which makes it difficult to distinguish
between them.

Instead use the `logfile_read` and `logfile_send` attributes to
separate the input and output logs. The pexpect inputs will now be
logged as inputs. The previous input logs that were made by ShellCommand
class will be logged as DEBUG and with repr() to log exactly how
it was called by other LAVA code.