tests: Cleanup and warn about unfinished SIGALRM
Because the lava_dispatcher code uses SIGALRM for timeouts it is
important that there shouldn't be any dangling SIGALRM scheduled
after running a test. Otherwise the next test might fail because
the SIGALRM will raise an exception scheduled in the previous test.

Also emit a warning so that offending tests can be fixed.