Preserve key order in yaml_safe_dump by default
Currently when the YAML is dumped all the dictionary keys will be
sorted. This can have unexpected consequences. For example, the
order of downloads will be sorted based on their names.

By having YAML preserve the key ordering the job definitions are
now more accurate as the downloads will now happen in their declaration
order.

This requires regenerating all pipeline_refs because the JobParser
is tightly coupled with YAML so it has to go through the
`yaml_safe_dump` -> `yaml_safe_load` cycle which would sort the
keys. The only test that needed adjustment was TestQemuNFS.test_qemu_nfs
where order of QEMU commands changed because of downloads order was
preserved. I tested QEMU and it seems to not care which in which
order was `-kernel` and `-initrd` called.