Avoid having to insert TestCase ids in to the logs before rendering
Currently every time the job logs have to be rendered the entire log
has to be iterated and for every TestCase in the logs its id has to
be inserted in to the logs. This is done to be able to create a URL
to the TestCase views.

However, this is completely unnecessary as TestCase view can be
accessed from its job id, testsuite name and name directly.
All those values are available from the logs themselves and no
modifications are required.

This also fixes the MOUNT_POINT setting not working with the
incremental logs.