Unify logs rendering code path
Right now there are essentially 2 different code paths that render
logs on the job details page. First is inside the django templates
and second one is inside javascript incremental logs.

However, this is suboptimal because it requires maintaining two
separate code paths. Instead merge it in to one code path inside
javascript.

For initial logs already present on the server use the `json_script`
filter provided by Django. It will convert the logs in to the JSON
object that can later be parsed by the javascript.

Move the action link list rendering to javascript meaning the
`_pipeline_actions.html` template is no longer needed.

This improves performance because there will be less load on the
server to render templates and reduces the required data transfer
because logs will be compressed in to compact JSON instead of HTML.