jlink boot: multiple flash added
JLink has been modified to allow flashing multiple binaries.
As a result, the use of the commands parameter in the job definition is now optional.
Additionally, the prompt parameter has been added to facilitate interactive testing.
This parameter is optional and should only be used when the flashed binary behaves like a shell.