lava_dispatcher: run job cleanup in a separate thread
After a job fails to clean up its temporary directory, lava worker may
take more than 60s to remove the job directory. When this happens, the
worker will be marked as offline by lava scheduler as the next ping to
the server is blocked.

This change runs job cleanup in separate threads managed by
ThreadPoolExecutor. Threads are reused and limited to 8.

Signed-off-by: Igor Ponomarev <igor.ponomarev@collabora.com>
Signed-off-by: Chase Qi <chase.qi@linaro.org>