linaro_django_xmlrpc: return None for close() method
Async API client aioxmlrpc <= v0.5 is based on aiohttp which needs an
explicit `close()` method to close the session. The latest aioxmlrpc is
based on httpx AsyncClient which always closes itself by default. But
the problem is that it sends the `close()` method when requested to the
server.

This change treats `close()` as a valid method and returns `None`. This
is mainly for unifying user-side use experience.

Signed-off-by: Chase Qi <chase.qi@linaro.org>