Make ShellSession.wait JobError message controllable
Use it for LinuxKernelMessages.parse_failures(). When a kernel panic
happens a board can reboot faster than the end of the kernel panic
could be printed. This will result the job error becoming "wait for prompt timed out"
which is very misleading.

Using this new parameter the error message will be "Kernel panic - not syncing"
even if matching the end of kernel panic fails.