Make JobError a kernel panic if it times out during end of panic search
If a device reboots too fast and does not print "end Kernel panic"
the final error would be "wait for prompt timed out" instead.

This is very misleading because the root cause is still kernel panic
so the job error message should reflect that.

To implement this once a "Kernel panic - not syncing" is detected
set the `halt` variable to "Kernel panic" and ignore any JobErrors
raised by `connection.wait`. This will result any timeouts raised
by `connection.wait` to be considered as a failure to match the end
of kernel error and the reader loop will exit. Once reader loop
exists the `halt` message will be picked up and a new JobError
will be raised which will contain kernel panic message.