Highlight actions on screen and auto-scroll pipeline table on job details page
Use `IntersectionObserver` to monitor which logs entries belonging
to witch actions are currently on screen. If logs from a certain
action are on the screen highlight this action in bold in the pipeline
table and scroll it to the middle of the table.

This makes it easy to know which action you are currently looking at.