tests: Restore working directory after lava_dispatcher/test_utils.py tests
Those tests involve calling the `os.chdir` but never restored it back
after the tests. This could potentially break certain tests if an
order randomization was enabled.

There is a context manager available at `lava_dispatcher.utils.contextmanager.chdir`
which can switch and then restore the working directory. Lets use it
both for its purpose and to test that it is working correctly.