Add type hints to lava_scheduler_app/logutils.py
This makes it possible to search where a methods of the Logs class
are used.

Unfortunately a lot of third party libraries are missing type hints
so it is not possible to completely check all subclasses. At least
the `LogsFilesystem` should be able to completely pass type checking
in the future.