Fix NoReverseMatch if TestCases with same names share TestSet
In `case.html` when a TestCases have the same names and same TestSet
the template tried to create a link to a view of all TestCases
belonging to same TestSet.

However, it tried to use the `lava.results.testcase` view but with
that view cannot make use of TestSet. Instead use `lava.results.testset`
which uses the same `case.html` template but will filter cases by
the passed TestSet.