device-type: base-fastboot: add minimal boot method
Some boards may have additional steps between a "fastboot boot" command
and making it to a Linux login prompt, for example the Qualcomm RB2 board
needs to trigger a lava command to set a USB mux.

Enable the minimal boot method for fastboot boards so this boot process
can be correctly modelled in LAVA jobs.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>