lava-worker: Fix Job.description returning bytes instead of str
The function was marked as `str` return type but actually used
`read_bytes()` and returned bytes. This did not cause any issues
because the only place the description used is in finished job
POST request encoded as `data=` that accepts both strings and bytes.
However, if in the future it could switch to using `json=` which
will raise exceptions if bytes are passed.