Skip favorite TestJobs when deleting jobs with lava-server manage
Add a new command line flags `--skip-favorite` and `--no-skip-favorite`
which control if the favorite jobs should be deleted. By default
favorite jobs would be skipped.

For example, this can be used to preserve certain TestJobs for
bug reports. Currently bug reports that point to a specific job
will lose information over time as the TestJobs are getting cleaned
up.