Cleanup man/conf.py
This file was generated in 2011 and contains a lot of unused options.

Import `lava_common.version` instead of calling a subprocess.
Remove all commented out options. They are probably outdated since 2011.
Remove LaTeX options.
Remove `index.rst` because it is unnecessary. Mark all man pages
as `:orphant:` because they don't need to be included in toctree.