Improve yaml_safe_dump and yaml_safe_load typing
Mask them under the `yaml.safe_dump` and `yaml.safe_load` for
type checkers since they expose similar API. To do this use the
existing `if TYPE_CHECKING` statement and only define the real
functions if it is not type checking. Add the `__all__` attribute
to make re-export explicit and not emit a warning.