lava_dispatcher: ensure containers are destroyed in job cleanup
Docker containers launched by `OptionalContainerAction` is removed by
the `docker.stop(self.action)` that is defined in the finally block of
`try ... finally ...` block. If the try block times out and the action
is killed by the action `on_timeout()` signal handler, it may leak
containers.

This change adds an action level cleanup method to ensure that all the
containers created by the action are destroyed before the job finishes.

Signed-off-by: Chase Qi <chase.qi@linaro.org>