lava_dispatcher: u-boot: Add support for raw initramfs images
Consider what happens if LAVA is handed a job that looks something like
this:

actions:
  - deploy:
      to: tftp
      timeout: {minutes: 3}
      kernel:
        url: "https://.../Image"
        type: image
      dtb:
        url: "http://.../board.dtb"
      ramdisk:
        url: "https://.../initramfs.cpio.gz"
  - boot:
      method: u-boot
      ...

Currently LAVA will prepare the overlay, inject it into the initramfs and
then wrap it with a header in the u-boot legacy image format. We are
increasingly seeing u-boot firmwares that disables support for legacy
image format so we'd like to be able to handle other initramfs formats,
initial starting with raw images.

As it happens LAVA already supports raw format for "initrd" images but not
for "ramdisk" images (and I think only the later supports overlays). Let's
extend the existing support for raw images by allowing device-type
templates to request a format for ramdisk images that aligns with the
bootloader capabilities.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>