device-types: base: Recognise kernel_comp_addr_r errors
If booti is used to boot a gzip compressed image then u-boot will
automatically try to decompress it. However this relies on two variables
which are often not set in the u-boot environment. If the variables are
not set then the boot will hang.

Add the error string to the default set so we can avoid timeouts if this
problem comes up.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>