lava_results_app: try to keep a partial result error_msg
Only the critical data is saved when `TestCase` metadata exceeds 4096.
This should work fine in most cases but may cause an empty `Job error: `
message on the UI. That is because the result `error_msg` is stripped
off in the metadata. Boot log parsing failure caused by a long kernel
panic error message is a good example for this.

This change tries to keep a partial error message to make it less
confusing on the UI. In this case, a job failure comment will also be
displayed just below the job error, telling user that the result
metadata is too long to show. This should be clear enough.

The length of the `error_msg` to keep isn't easy to decide. The same as
the job failure comment, 256 is used here.

Signed-off-by: Chase Qi <chase.qi@linaro.org>