Fix navbar anchor problem using CSS instead of Javascript
The problem was that because the navbar is fixed position scrolling
to an element resulted in it being under the navbar.

See upstream issue: https://github.com/twbs/bootstrap/issues/1768

The old solution was using the javascript which resulted in jitter
and was broken if same anchor was clicked twice.

New solution uses the CSS `scroll-margin-top` property which is
widely supported in all major browser for the last 5 years.