lava-worker: Batch start jobs
When starting a new job worker has to do 2 HTTP requests to the
server. Instead of waiting in sequence to start jobs use asyncio
to start them in parallel.

Use `asyncio.gather` to run the job start coroutines. By default
it will raise the first exception if any of the coroutines raise
an exception. Cancel the remaining coroutines by canelling gather.
This most closely replicates the current behavior where the main
loop aborts if any of the job raises an exception during start.