lava-worker: Batch start jobs
When starting a new job worker has to do 2 HTTP requests to the
server. Instead of waiting in sequence to start jobs use asyncio
to start them in parallel.

Use `asyncio.TaskGroup` to run the job start coroutines.
It closely replicates the current behavior where the main
loop aborts if any of the job raises an exception during start.

Signed-off-by: Igor Ponomarev <igor.ponomarev@collabora.com>