lava_common: report an error when a log line is too large to upload
Nginx sets the request body size limit to 1m by default. Logline that
is close to `10**6` in length plus the extra keys we add in the request
body could be larger than the 1m limit. In this case, the logline will
never be uploaded and the job will hang there. Boot log parsing result
line with many trace logs inside is good example of this issue.

Reducing the line size to `10**5` should make it work with nginx's
default limit. apache2 doesn't set a limitation by default.

In case a signal line is still too large to upload, report a failure
and print an error message to kill the job.

Signed-off-by: Chase Qi <chase.qi@linaro.org>