dispatcher: Add a rate limiting for download process updates
Even with small downloads (e.g. 5 MB) which take less than 0.1 seconds to download there is a progress update of the download every 5 % step. That's way too much and just clogging the logs. So only report a progress if it takes longer than .1 seconds and report progress only every .1 second