Download images in job definition order
Thanks to the fc4f68a587ff4b2c548b3048bdd24ed5b6c08c0f the key order
in YAML that lava-worker downloads is now preserved. This means
it is no longer necessary to sort the job downloads and they
can happen in the same order as they are defined in job.