Improve scheduler triggering when a job is finished
When a job is finished, the Device state is set to Idle then the TestJob state
is updated to Finished inside a common transaction.
The corresponding events will be sent immediately (Device then TestJob). The
scheduler will use the first event (Device state change) and try to schedule on
the first event. However, the transaction may not be committed yet.

Allow the scheduler to also use the TestJob state to trigger scheduling which
is sent right before committing the transaction.

Also remove filtering on available device-types as this can force the scheduler
to only consider a subset of the device-types.