lava-run: Print LAVABug traceback
Currently if the job finishes because of LAVABug only its message
will be saved and the traceback would be discarded. This makes it
harder to debug the actual LAVA bugs. This is because LAVABug is
a subclass of LAVAError the code path that prints tracebacks is
not used

Add a codepath to print LAVABug traceback to the LAVAError handler.