lava_dispatcher: check if device is shared to docker container correctly
Lab engineer spotted the below issue on lava docker worker host.
Restarting `lava-dispatcher-host.service` solved the issue. The reason
behind is to be investigated when we have the debugging environment.

```
systemd-udevd[611768]: 2-1.1.1.1:1.0: Spawned process
'/usr/bin/lava-dispatcher-host
--debug-log=/var/log/lava-dispatcher-host.log devices share --remote
2-1.1.1.1:1.0 --vendor-id=18d1 --product-id=d00d' [611769] is taking
longer than 59s to complete
```

This change checks if the device is shared correctly before running
commands in the docker container. It retries the sharing 3 times with
60s delay. This will log the sharing error and kill the job earlier
if all the retries fail.

Signed-off-by: Chase Qi <chase.qi@linaro.org>