Sentry: Add SENTRY_TRACES_SAMPLE_RATE for error and performance tracking
This is required in the SDK to allow sentry to be used for performance
metrics enablement.

To send a representative sample of your errors to Sentry, set the
sample_rate option in your SDK configuration to a number between 0
(0% of errors sent) and 1 (100% of errors sent). This is a static
rate, which will apply equally to all errors.

Signed-off-by: Benjamin Copeland <ben.copeland@linaro.org>