lava_dispatcher: Remove `Accept-Encoding: ""
Depends on MR[1], which changes LAVA behaviour to explicitly disabled automatic
content decoding by setting `decode_content=False` in `requests`.
This ensures that the downloaded content is received as-is, without any unexpected
decompression.

Removed the `Accept-Encoding: ""` header to prevent reliance on inconsistent
server-side behavior when handling compressed files.

This commit, along with [1], and setting Content-Length will ensure LAVA
will download the file, and push the problem down to user level. This
removes the ability to be reliant on requests library.

1: https://gitlab.com/lava/lava/-/merge_requests/2636

Signed-off-by: Benjamin Copeland <ben.copeland@linaro.org>