Use which helper inside rpcinfo_nfs function
Currently in the PersistentNFSOverlay `validate()` has to manually
call the `which()` function to raise InfrastructureError if the
`rpcinfo` is installed. If this is called beforehand a LAVABug would
be raised.

Instead of having manually call the `which("rpcinfo")` move this
call inside the `rpcinfo_nfs` function.

Use `subprocess.run` instead of `subprocess.Popen` as it will
automatically cleanup subprocess. Add a small timeout so that
subprocess does not block the job execution.

This also makes it easier to mock in unittests as only one function
call needs to be patched.