Use urllib3's enforce_content_length instead of extra HEAD request
Instead of making an extra HEAD request to record the size of the
download use the `enforce_content_length` option that urllib3 provides
to achieve same goal of detecting truncated downloads. The urllib3
will raise an exception if the download did not complete.

To make the download progress be logged set the size from the GET
request and reset it back to -1 as soon as download ends to make
DownloadHandler not check the size again. Separate all download
progress logic in to a new class and lazy log the download size
as at will only be known once the HTTP request starts.

Use correct MiB units in logging as the download sizes are divided
by 2**20 not 10**6.

This also makes LAVA compatible with the HTTP servers that cannot
send their uncompressed sizes such as S3.