Only compute hashes of downloaded files that will be checked
Computing cryptographic hashes like SHA-512 is very expensive.
Currently the DownloadHandler will compute the MD5, SHA-256 and
SHA-512 every time the file is downloaded.

However, the file hashes will only be checked if the job definition
provided the hex values to compare to. There is no point in calculating
the hash digests if they will never be checked.

Also remove the hash namespace data and test case results. The hash
compare function will already put the hash values in to the logs and
test case results.