lava_scheduler_app: fix job page jumping on mouse wheel scrolling
When user is on the top of a job page, the job page keeps on jumping
back to the top when user tries to use the mouse wheel to scroll down.

It doesn't happen for every job, when it happens, it is the
`action_table_entry.scrollIntoView` action competing with user input.

Running the scroll action only when the action table is floating solves
the issue.

Signed-off-by: Chase Qi <chase.qi@linaro.org>
Signed-off-by: Igor Ponomarev <igor.ponomarev@collabora.com>