lava_dispatcher: check error message to ensure job not found on server
Certain HTTP proxies like Traefik can send 404 if they consider the
backend to be malfunctioning. This can result in the same job being
restarted multiple times.

This change checks both the returned error code and error message to
ensure the job is deleted on the lava server before deleting it on the
lava worker.

Signed-off-by: Chase Qi <chase.qi@linaro.org>