postinst: Use psycopg2 in is_pg_available function
Because `postgresql-client` is no longer a required dependency
the `pg_isready` might not be installed.

To check if PostgreSQL is available use the same mechanism to
connect to PostgreSQL server as the `create_database` function uses.
Switch to `postgres` user and try opening the connection. If no
error gets raised that means the PostgreSQL is available.