Remove parsing undocumented lava_server/.env file for prod settings
The issue is that the `lava_server` package folder is usually located
in `/usr/lib/python3/dist-packages` which is reserved for the system
installed packages. Usually the system settings go in to the `/etc`
folder.

This file was also completely undocumented and required ugly
warnings silence because before version 0.7.0 django-environ
would raise a warning if this file was not present. Only Debian
13 ships the django-environ that does not raise warnings.