lava_dispatcher: save "downloads://*" file to unique path by default
The same `flasher` command passes with `http(s)://` schema but may fail
with `downloads://` schema. This is because the download action for the
former schema downloads file to a unique path but uses the same
directory to download files for the latter schema.

This change unify the default behavior cross `http(s)://`, `file://` and
`downloads://` schemas and it provides the `uniquify` parameter for
overwriting.

The downloaded file `self.path` is only used internally, so this change
shouldn't introduce regressions.

Signed-off-by: Chase Qi <chase.qi@linaro.org>