lava_dispatcher: fix frequent worker offline
After 2024.09, frequent online -> offline -> online records are
observed on most workers. It is a regression introduced by the commit
below.

be0f40cbf7c7e95fb03bd83507a84ccaf43d9baf

The commit switched to aiohttp for more efficient connection reuse.
The problem is that aiohttp client session manages connection with a
default `keepalive_timeout = 15.0`. When no job is being processed, the
lava worker pings the lava server every 20s. In this case, the existing
connection will be closed and not be reused.

This change increases the client session `keepalive_timeout` and also
sends explicit header to the server for connection reuse.

Signed-off-by: Chase Qi <chase.qi@linaro.org>