depthcharge: Fix not being able to override boot commands
The issue was `DepthchargeCommandOverlay.validate` always assumed
that `commands` parameter can only be a string. However, the job
definition allows overriding the boot commands defined by a device
dictionary by passing a list to the `commands` parameter. This caused
the `DepthchargeCommandOverlay` to raise a `TypeError: unhashable type: 'list'`
exception.

Add a check to the `validate` method to only fetch the `cmdline`
from the device dictionary if the `commands` is a single string.
If a list gets passed assume that the `cmdline` is empty.