lava_dispatcher: use ping interval for api request timeout
Because the current client session timeout is set to 10 minutes by
default, in case of any network issues, it could take a long time to
create a connection or to timeout. It will hold the main loop and leads
to worker offline on the server side.

Setting keepalive timeout to a larger value than the web worker timeout
and api request timeout to ping interval should be more robust. It
reuses the connection as long as possible and times out in time for the
next loop to keep the worker alive.

Signed-off-by: Chase Qi <chase.qi@linaro.org>