black: Disable magic comma and reformat all Python files
Imagine that originally you had a very long line that got split
by black in to multiple lines with a trailing commas:

```python
[
   "veeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeery_loooooooooooooong1",
   "veeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeery_loooooooooooooong2",
]
```

However, at some point the long items got replaced with short ones:

```python
[
   "short",
   "short",
]
```

The issue is because of trailing commas black will never try to
condence the declaration back to single line. However, it has
a flag `--skip-magic-trailing-comma` which makes it ignore the
trailing commas. This will allow the declaration to be formatted
back to single line:

```python
["short", "short"]
```

Enable this mode for the CI and reformat all sources.