lava_dispatcher: run avh image upload in a separate thread
Once the avh image uploading started, it blocks job to timeout or to be
cancelled. And after the job is terminated eventually, it leaves a
process behind.

This change moves the uploading to a separate thread in daemon mode so
that it can be closed by the main thread or the OS.

Signed-off-by: Chase Qi <chase.qi@linaro.org>