lava_dispatcher: clone test repository with submodules
When cloning a test-definitions repository use --recurse-submodules to
also bring any potential submodules to the resulting directory.

To maintain backward compatibility recursive cloning is disabled by
default. To enable the feature additional key in the test/definitions is
added: "recursive". When set to true LAVA will attempt to clone with
--recurse-submodules

Signed-off-by: Milosz Wasilewski <quic_mwasilew@quicinc.com>