codespell: Use MR pipeline vars to fetch and list of commits
The previous script relied on `master` branch to be up to date
in the repository from which the merge request was created
otherwise the list would dump all known commits to the codespell
and fail.

This new script uses the environment variables gitlab provides
to fetch the targeted branch and build a list of commits.

This means the **`master` branch no longer needs to be up to date**
which was a common issue for contributors. Also the new script
will work accurately for MRs tareting branches other than default
branch therefore the old script always compared against
`CI_DEFAULT_BRANCH`.

`CI_MERGE_REQUEST_PROJECT_URL` contains the URL to the targeted
repository. It might use the `https` protocol so make sure the
`ca-certificates` is installed otherwise `https` protocol might
not work. Add the `CI_MERGE_REQUEST_PROJECT_URL` repository as
the `upstream` remote.

The `CI_MERGE_REQUEST_TARGET_BRANCH_NAME` and `CI_MERGE_REQUEST_SOURCE_BRANCH_NAME`
contain the names of the target and source branch. After fetching
the target branch from merge request project repository the
`..` can be used to build the exact list of commits.

This new pipeline stage uses rules to only start when the merge
request is created because the variables used are only availabe
in merge request pipelines. This requires `except:` to be `null`
or otherwise there will be conflict between rules based triggers
and except ones.