Do not rely on compression auto naming (fix .zst vs .zstd)
The previous code delegated naming the compressed file to the called
binary and then assumed it was called `{filename}.{compression_name}`.
This worked for `gz` but Zstandard encountered the issue. LAVA expected
the `.zstd` suffix but the `zstd` command added `.zst`. This resulted
in compressed files being ignored.

Fix this issue by not letting the compression binaries pick names.
Instead attach the compression subprocess stdout to the correct file
opened by LAVA. This means the final file name will be fully controlled
by LAVA.

Add a unit test to verify that all known compression algorithms
will work with stdout being attached to the final file.

Add compression binaries packages to unit tests requirements as
they might not be installed on the system.